requirements.txt
torch==2.5.1+cu124
matplotlib==3.10.0
tqdm==4.67.1
numpy==1.26.3
streamlit==1.41.1
pillow==10.2.0
onnxruntime==1.20.1
torchvision==0.20.1+cu124